+++ framework-crates/objc2-metal/examples/triangle.rs
@@ -3,17 +3,16 @@
 
 use core::{cell::OnceCell, ptr::NonNull};
 
-use objc2::{
-    declare_class, msg_send_id, mutability::MainThreadOnly, rc::Retained, runtime::ProtocolObject,
-    ClassType, DeclaredClass,
-};
+use objc2::mutability::{IsMainThreadOnly, MainThreadOnly};
+use objc2::rc::Retained;
+use objc2::runtime::ProtocolObject;
+use objc2::{declare_class, msg_send_id, ClassType, DeclaredClass, MainThreadMarker};
 use objc2_app_kit::{
     NSApplication, NSApplicationActivationPolicy, NSApplicationDelegate, NSBackingStoreType,
     NSWindow, NSWindowStyleMask,
 };
 use objc2_foundation::{
-    ns_string, MainThreadMarker, NSDate, NSNotification, NSObject, NSObjectProtocol, NSPoint,
-    NSRect, NSSize,
+    ns_string, NSDate, NSNotification, NSObject, NSObjectProtocol, NSPoint, NSRect, NSSize,
 };
 use objc2_metal::{
     MTLCommandBuffer, MTLCommandEncoder, MTLCommandQueue, MTLCreateSystemDefaultDevice, MTLDevice,
@@ -86,7 +85,7 @@ declare_class!(
         #[method(applicationDidFinishLaunching:)]
         #[allow(non_snake_case)]
         unsafe fn applicationDidFinishLaunching(&self, _notification: &NSNotification) {
-            let mtm = MainThreadMarker::from(self);
+            let mtm = self.mtm();
             // create the app window
             let window = {
                 let content_rect = NSRect::new(NSPoint::new(0., 0.), NSSize::new(768., 768.));
@@ -97,7 +96,7 @@ declare_class!(
                 let flag = false;
                 unsafe {
                     NSWindow::initWithContentRect_styleMask_backing_defer(
-                        mtm.alloc(),
+                        NSWindow::alloc_main_thread(mtm),
                         content_rect,
                         style,
                         backing_store_type,
@@ -120,7 +119,7 @@ declare_class!(
             // create the metal view
             let mtk_view = {
                 let frame_rect = window.frame();
-                unsafe { MTKView::initWithFrame_device(mtm.alloc(), frame_rect, Some(&device)) }
+                unsafe { MTKView::initWithFrame_device(MTKView::alloc_main_thread(mtm), frame_rect, Some(&device)) }
             };
 
             // create the pipeline descriptor
@@ -281,7 +280,7 @@ declare_class!(
 
 impl Delegate {
     fn new(mtm: MainThreadMarker) -> Retained<Self> {
-        let this = mtm.alloc();
+        let this = Self::alloc_main_thread(mtm);
         let this = this.set_ivars(Ivars {
             start_date: unsafe { NSDate::now() },
             command_queue: OnceCell::default(),