+++ crates/objc2/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 * Added `AnyClass::is_metaclass`.
+* Added `MainThreadMarker` from `objc2-foundation`.
+* Added `ClassType::alloc_main_thread`.
+* Added `IsMainThreadOnly::mtm`.
 
 
 ## 0.5.2 - 2024-05-21